polymarket-velocity-bot/requirements.txt
py-clob-client==0.34.6
python-binance==1.0.19
websockets==14.1
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
torch==2.6.0
pyyaml==6.0.2
asyncio
loguru
ccxt
scipy
